feat: add userId handling for shopping list operations and implement anonymous user ID generation

+++ src/app/api/shoplist/[id]/route.ts
@@ -8,10 +8,16 @@ type Context = {
 
 export async function DELETE(request: NextRequest, context: Context) {
   const { id } = await context.params;
+  const { searchParams } = new URL(request.url);
+  const userId = searchParams.get("userId");
+
+  if (!userId) {
+    return NextResponse.json({ error: "Missing userId" }, { status: 400 });
+  }
 
   try {
     await prisma.shopping_list.delete({
-      where: { id: Number(id) },
+      where: { id: Number(id), userId },
     });
     return NextResponse.json({ message: "Item Deleted" });
   } catch (error) {
@@ -24,11 +30,18 @@ export async function DELETE(request: NextRequest, context: Context) {
 
 export async function PUT(request: NextRequest, context: Context) {
   const { id } = await context.params;
+  const { searchParams } = new URL(request.url);
+  const userId = searchParams.get("userId");
+
+  if (!userId) {
+    return NextResponse.json({ error: "Missing userId" }, { status: 400 });
+  }
+
   const body = await request.json();
 
   try {
     const updatedItem = await prisma.shopping_list.update({
-      where: { id: Number(id) },
+      where: { id: Number(id), userId },
       data: {
         item: body.item,
         priority: body.priority,
